fix: module resolution inside worker scripts in app builds
