+++ packages/bruno-electron/src/bru/workerScripts/bruToJsonV2Worker.js
@@ -1,0 +1,13 @@
+const { workerData, parentPort } = require('worker_threads');
+const {
+  bruToJsonV2,
+} = require('@usebruno/lang');
+
+try {
+  const bru = workerData;
+  const json = bruToJsonV2(bru);
+  parentPort.postMessage(json);
+}
+catch(error) {
+  console.error(error);
+}
